Add PagingState
Test: ./gradlew paging:p-c:test

Contains full snapshot of all paging state to be presented by adapter.

PageEvents can be dispatched to a PagingState.Producer, which will
allow us to transform a Flow<PageEvent> -> Flow<PagingState>.

Note that PagingState does not currently come with an easy way to diff
- currently always have to use DiffUtil to compare even trivial
changes (such as page added), but this will be optimized. Once these
optimizations are in place, tests in PagingStateTest can assert on
insert/remove/changed events that an adapter would receive.

Additionally in this CL, starting to standardize on naming of
placeholdersStart/placeholdersEnd (vs 'nulls', or
'before'/'after'). Will clean up further in follow up.

Change-Id: I7543954256dd07c11cc0a4a00407ccea83790ed4