Add LayoutTag modifier
The CL adds a LayoutTag modifier which can be used to add parent data
representing a tag associated to a child. For example, this can be used
to associate composed children to measurables. The modifier will also be
used to associate ConstraintLayout children to their correspondent
representation in the ConstraintSet. LayoutTag is a ParentDataModifier,
a new type of modifiers which only modify parent data.

Bug: 145197157
Bug: 142722751
Test: ran androidx.ui.core.test.TagTest
Change-Id: I69fe48a28c3a5280fc0dc931976254ab557f440b