Add DisplayCompat to get physical display size
Create DisplayCompat static shim with nested ModeCompat classes. This is necessary to get the display size for 4k displays, because current atv
devices do not correctly handle / identify / add modes which results in
no way for apps to identify the 4k capability of a display.

The intended way of finding out if a display is capable of displaying 4k
content is to iterate through display.getSupportedModes() and check if a
mode with 4k resolution exists. The ModeCompat class has an additional
field isNative with which a client can identify that this is a native
mode. The DisplayCompat class adds an additional ModeCompat to the
supported modes of the wrapped Display object. This way the clients can,
as intended, iterate through the displayCompat.getSupportedModes() and find the
highest native resolution to verify if the display is 4k capable.

The issue will be fixed by adding CTS tests and enforcing the correct behaviour
in future Android versions.

Bug: 138275089
Test: DisplayCompatTest
Change-Id: Id927ff58cd697381ed1de16094cee3b5790c9550