Add WindowInsetsCompat.Builder to core
This backports the new WindowInsets.Builder class
to the AndroidX core library. Since there is no
public way to create an WindowInsets instance on
API < 29, we need to use reflection on those platforms.

Both of the methods we use via reflection are available
in AOSP from SDK 20, so at least one of them *should*
be on available on devices. If not, the impl fallbacks
back to an empty instance.

While there, I also tidied up the nullability annotations,
start caching the generated Insets instances, and updated
and some docs on WindowInsetsCompat and Insets.

Had to pin ViewPager back to using customview:1.0.0, otherwise
it pulls in the new changes and -werror breaks the build because
of using (newly) deprecated APIs. The last ViewPager release
was v1.0.0 so won't affect any public releases.

BUG: 145709065
Test: ./gradlew core:core:connectedCheck

Change-Id: I1286afcf9a09d28ad10df636733149e6ac4126bd