Fix black screen issue of CoreTestApp below API 23
There will be the black screen of Preview when reusing DeferrableSurface.
The problem is the behavior of TextureView below API 23. It releases SurfaceTexture
when activity onStop. Hence,request to close and recreate DeferrableSurface below
API 23.

Apply this fix, it is okay to return true or false in onSurfaceTextureDestroyed
of TextureView.SurfaceTextureListener.

Bug: 145725334
Test: manual device test && ./gradlew bOS &&
./gradlew camera:camera-core:connectedAndroidTest &&
./gradlew camera:camera-camera2:connectedAndroidTest &&
./gradlew camera:camera-core:test &&
./gradlew camera:camera-camera2:test &&
./gradlew camera:integration-tests:camera-testapp-core:connectedAndroidTest

Change-Id: I949ab49587541353e1fe7684017f6069b4f38a45