Add Generate All Release Notes Task
generateReleaseNotes.py queries the jetpad database for the release
information of a specific release.  It then creates a json file that
is the serialized info that release, and passes it to the gradle
task ./gradlew generateReleaseNotes

The GenerateAllReleaseNotes task is an aggregator task that depends on
the GenerateArtifactReleaseNotes task.  The
generateArtifactReleaseNotesTask
creates the release notes for each individual artifact, if that
artifact was specified in the json file from Jetpad.

GenerateAllReleaseNotes then iterates over all the output file from
the GenerateArtifactReleaseNotes task and aggregates them by groupId
and also creates a master androidx_aggregate_release_notes.txt file
that contains all the release notes for this specific release.

It can be run like `./gradlew generateReleaseNotes -PartifactToCommitMap=Myjsonfile.json`

Test: cd development/release-notes/ && ./generateReleaseNotes.py 1570640400000 && subl ../../../../out/dist/release-notes/androidx_aggregate_release_notes.txt
Test: cd development/release-notes/ && ./generateReleaseNotes.py 1575482400000 && subl ../../../../out/dist/release-notes/androidx_aggregate_release_notes.txt
Test: gw tasks --group Documentation | grep generateReleaseNotes
Test: gw bOS
Test: gw buildSrc-tests:test

Change-Id: Iad42e4e7203ae38f012501b6539eb697d008df0c