Call transitionToIdle callback in Espresso.
It turned out that several of our tests run a bit longer than needed
(e.g. 5-15 sec instead of 100s' of ms). It turned out that even though
the Espresso Idling Resource API looks like PollingCheck it actually
expects us to call the callback once we get idle. If that does not
happen Espresso goes to "sleep" for 5 seconds which is essentially first
part of failure mode. After that it tries to call us again.

Bug: b/136262330
Test: Still pass (waiting for presubmit to verify better times).
Change-Id: Idea1bdbd83b83050323e553df3c7c06384d48ed5