Fix DeferrableSurfacesTest fail
Issue root cause:
(*) The fake ListenableFuture which is used in this test  might be completed
when its completer is finalized (gc).
(*) The DeferrableSurfacesTest might fail due to the fake ListenableFuture
is completed earlier than the test is verified.

Solution:
We can keep the completer instance to avoid the garbage
collection, and complete the future after the test is finished.

Test: ./gradlew bOS &&
./gradlew camera:camera-core:connectedAndroidTest &&
./gradlew camera:camera-camera2:connectedAndroidTest &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:camera-core:test &&
./gradlew camera:camera-camera2:test &&
./gradlew camera:camera-extensions:test &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:integration-tests:camera-testapp-extensions:connectedAndroidTest

Change-Id: I0786b7346c82da1363711830460c2f4616255ed9