Fix Preview black screen on Legacy camera and API > 23 devices.
Root Cause:
- When unbind all online usecases, it invokes configAndClose. If
it did not close camera in sequence, but opened CaptureSession from
other requests. The Camera state will be changed to open, but the opened
CaptureSession will be reset in the future callback of openDummyCaptureSession

Solution:
onSuccess/onFailure part for the callback of future:
Don't need to invoke resetCaptureSession again. There is the releaseSession
later to make sure close the dummy session and then close camera in corret state.

- Test: After aosp/1198183, the test of openNewCaptureSessionImmediateBeforePreviousCaptureSessionClosed
can check the case, the 2nd assertTrue(mSessionStateCallback.waitForOnConfigured(1)).

Bug: 146591609, 147402661
Test: Verify on Nokia 1 and Pixel API 26 emulator && Pass
      ./gradlew :camera:camera-camera2:connectedCheck &&
      ./gradlew :camera:camera-core:connectedCheck &&
      ./gradlew :camera:camera-view:connectedCheck &&
      ./gradlew :camera:camera-extensions:connectedCheck &&
      ./gradlew :camera:camera-lifecycle:connectedCheck

Change-Id: Ia7a35e94aa8a48be2cc026bbafd1f0d5f37a1d3d