Cleanup Serialization's Reserved API in Java
Previously, the Java projection of Reserved.ids did not properly default
to an empty int[]. While this worked as expected in Kotlin, it made it
un-ergonomic to reserve ID ranges or names without individual IDs in
Java.

Additonally, I had used @JvmName("value") on ids to allow using Reserved
without keywords in Java, but this proved to be less ergonomic, so I
removed it, requiring it to be used as a keyword attribute.

Test: ./gradlew :serialization:serialization-annotation:build
Bug: 144724751
Change-Id: I0be18f7b41106d7a8eb20ecf3c4b8b5016c888c1