Serialization schema compilation for Java enums
This CL adds a the ability to SchemaCompilationStep to extract
@EnumValue and @Reserved annotations from a processing environment and
render them to schema types. Kotlin metadata handling will be added for
enums later, after support for messages and services is complete.

It also includes a set of extension methods that make the annotation
processing API easier to work with in idiomatic Kotlin.

Validation is quick and dirty right now, and will need to be fleshed
out in order to scale with the code base.

Test: ./gradlew :serialization:serialization-compiler:test
Bug: 144724751
Change-Id: Icf6429730a6e39849b7e35bc28e5bb9ef782c787