Reuse RowScope and ColumnScope
The CL makes RowScope and ColumnScope reusable across instances and
recompositions, as they are used just to limit modifiers visibility and
do not have internal state. We can make them actual singletons if we
decide that it is okey for their content to be more easily acessible
outside the scope.

Test: ran FlexTest
Change-Id: Ib8b71c24a4b12ff0d953f568dcbb6ddb20e11c7c