Improve AndroidOwner assertions and fix requesting remeasure of the root during the measure pass
I made assertions in AndroidOwner a bit more readable and run our samples containing the complex screens and interactions with this flag enabled. This allows me to find few more cases leading to the inconsistent layout states. This cl improves assertions logic and fixes found issues. All this assertions and fixes should decrease amount of crashes happening because of drawing happening on the layoutnode which is not measured or laid out.

Test: manually on our samples with the assertion flag enabled plus a new test
Bug: 148278782
Change-Id: Ifacb5e1775411c46b3eb6443b72eef606ce78ae5