Fix bounding box calculation
Union with emptyBox should result in the other value being returned.
This fixes the calculation of boundary boxes.
The CL also changes how the absolute location of the LayoutNode is
obtained. Now we use positionRelativeToRoot so it accounts for padding

Updated the SlotTree Group generation to use IntPx for boundaries.

Bug: 139476297
Test: New test passes
Change-Id: I48235e393d0322d81f056afd4276d19b1377357e