Ensure that progress destinations are restored
Progress destinations get dynamically added to
the graph, which works fine for the first time the
graph is created, but fails if the progress destination
is on the back stack when the state is saved and
restored.

By keeping track of what DynamicNavGraph instances
don't have a default progress destination set, we
can ensure that when onRestoreState does occur, we
eagerly add the default progress destinations to
avoid "IllegalStateException: unknown destination during
restore" issues.

Test: Ran the sample issue in the bug
BUG: 148316337
Change-Id: I14ebfc0b3504e8ed7629f67268aa845bb96af708