Ensure calling onBackPressed() does not fail because of savedState
aosp/953431 introduced a fallback Runnable for OnBackPressedDispatcher
for when there are no enabled callbacks. On APIs below 26, if the
Activity state is saved, and onBackPressed() is called, there is a call
to the platform version of FragmentManager that throws an
IllegalStateException.

This change catches the known error regarding the saved state, and
throws all others.

Test: Added testCallOnBackPressedWhenStopped
BUG: 146290338
Change-Id: I0c24fc80944f7eaa6a79e0d23614e8a1c39d42a2