BUGFIX: fix failed saveToUri() tests in MobileHarness
When ImageCapture saves image to URI, if the default picture folder is deleted, it will throw a FileNotFoundException. This is the default behavior for Android. See MediaProvider#ensureDefaultFolders. Although we are unsure if this is the right behavior for CameraX, at least we can fix the tests by creating the folder.

Also fixed an issue that if ImageCapture will crash if MediaStore fails to create the URL.

Bug: 149177060
Test: manual test and ./gradlew bOS
Change-Id: I73718193c402e627c43db0e993ced722740940c9