Fix missing header/footer separators
Fixes: 150450889
Test: ./gradlew paging:p-c:test paging:p-ru:cC

Full rewrite of separator logic, to now display headers/footers (aka
'terminal' separators). Now we track all the necessary state to
support tracking load state and terminal separator insertion even when
events are empty.

Rewrote the filterEmptyPages behavior to avoid crashes in corner
cases, using new 'scan' operator. This operator is intended to be
basis for 'dedupe' behavior in the future as well.

This change also starts the process of moving transformation
operations out of the PageEvents themselves (for now, just filtering
empty pages). While `when() is ->` is less efficient than a virtual
function, the transformation logic is too complex to be sprinkled all
over the file.

Change-Id: I76ad6f965ae0b4e8df72b594feec355ca19b63f3