Allow findReferenceChild to find 0px child on edge
When a child is zero pixels high/wide and is at the edge of
RecyclerView's viewport, it is considered out of bounds because it is
"touching" the viewport from outside.

However, before r.android.com/1221510 this was considered within bounds.
Because neither the old nor the new behavior was wrong, reinstate the
old behavior to avoid regressions.

Bug: 148869110
Bug: 148453703
Test: ./gradlew recyclerview:recyclerview:cC \
        -Pandroid.testInstrumentationRunnerArguments.class=\
        androidx.recyclerview.widget.LinearLayoutManagerFindZeroPxReferenceChildTest

Change-Id: I6aaa117682bcda3aa5418069df53fcc7e3be6f2a