Enable animations for updateHiddenView test
This test tries to change an item while it is animating.
Unfortunately it'll flake when duration scale is 0 as there
is a change that update will happen after animation ends in
case instrumentation thread's enqueued update is scheduled
after the next UI frame.

In theory we can fix this if we had a test controlled item
animator. We dont so this is the easiest fix to unblock the
fast test runner.

I've run the test 100 times so pretty comfortable w/ this
fix.

Bug: 149527617
Test: for n in {1..100}; do ./gradlew recyclerview:recyclerview:connectedAndroidTest\
 --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.recyclerview.widget.RecyclerViewLayoutTest#updateHiddenView 2>&1\
 | grep "updateHiddenView\[" ; done
Change-Id: Ic8fb4524ed58563f447a0e98b03c9a0eab977983