Add layout direction to Layout node.
The layout direction will be passed from parent to children during the measure. Layout direction is exposed to measure block and intrinsic measure block (instead of the Constraint object as in aosp/1236391). Layout modifier methods also get LD as a parameter.
Added LayoutDirectiionModifier (named this way at the moment to not clash with the LayoutDirection enum).

Bug: 147608318
Test: tests passed, added tests for modifier
Relnote: "Made the layout direction be propagated from parent layout node to children. Added layout direction modifier."
Change-Id: I3d9559ddec464850d22466793975b41757e0224e