Fix issue where Python GitClient assumes subProjectDir is the root
For git logs, a git subdirectory project directory must be passed.
However, if you incorrectly pass a subdirectory with a leading
slash, the os.path.join() assumes that subdirectory exists in the
root directory.  That should never be the case, so in this case,
git log simply fails because we need the subdirectory, not an
absolute file path.

Test: ./generateReleaseNotes.py 1583344800000
Test: ./testReleaseNotes.py
Change-Id: I882dbe0c5aa05938cb95f317e099720a9d7e8433