Fix crash happening sometimes after calling disposeComposition()
The issue is when we dispose the whole tree we iterate through the tree and detach all the children. But we need to first detach our current node(set owner to null) before calling child.detach() as otherwise the children will call onRequestMeasure/invalidate on a parent. We can avoid this work as there is no reasons to call onRequestMeasure on the node which is also detaching. This was also making possible to request remeasure or invalidation on the already detached node in some cases causing crashes.

Fixes: 150728822
Test: manually, tests are passing, added one new test
Change-Id: Iebe0b8e24eba36916fcdf1a2ffdf20011f23ee61