Fix duplicate items handling in diff util
This CL fixes a bug in DiffUtil where it would get confused when
and item in the original list is duplicated, by shortcircuting the
path between the duplicate instances.

This CL fixes that bug by "unfortunately" re-implementing the
snake finding and path finding logic. I just couldn't figure out
what was wrong in the previous implementation and it was just too
hard to read with a billion offsets etc everywhere.

This implementation is easier to read, more true to the algorithm
in the paper at the cost of some more function calls.

On a pixel 3 (where allocations matter less), it is consistently
faster ~10-20% except when large item chunks move where it is
5% slower.
see measurements: https://paste.googleplex.com/6430892597706752

Bug: 123376278
Test: DiffUtilTest with duplicate cases

Change-Id: I94f64db7162c0107f9af512cdd55c0d39b44cabf