Removes ui-material-icons-extended dependency from demos
Compiling this module from source in demos hurts productivity as it takes a few minutes to recompile whenever a lower level module is changed.
This will be alleviated in the future when we split up our modules better, as we only need a very small amount of API in this module, but for now just copying and pasting the icons into the demos and taking the module out of demo compilation will help developer productivity.
In the future we also have the potential to just rely on prebuilts when we release stable versions.

Also fixes minor issue where the back icon was not showing, due to not observing the backstack.

Test: ./gradlew ui:i-t:demos:installDebug
Change-Id: Ib595638fd5d19418a6a4ffb0066fdec0b55c27bd