Fix crash when retrying REFRESH loads.
Instead of updating a callback which leads to racy scenarios, we instead
fan-out a retry signal from the adapter to all Pagers, so that signals
cannot be incorrectly consumed by an invalidated Pager before the
callback updates to point to the next generation Pager.

To prevent BroadcastChannel from buffering retry signals, consumption of
retry signals within Pager are wrapped behind a RendezvousChannel, which
should never block as retry signals should always be offered.

Fixes: 151323100
Fixes: 149525661
Test: ./gradlew paging:paging-common:test
Change-Id: Idb327a0e66dbcd96a2200329c5c404a18f49a010