Ensure state of dialogfragment without view is restored
In the case of creating a DialogFragment that does not have a view, if
onCreateDialog returns an AlertDialog and the process is restarted (such
as in the case of a config change) the previous state is not properly
restored. The reason for this is that we now save and restore state in
the onViewStateRestored callback, which does not happen for dialog
fragments without a view.

We need to make sure to save and restore state when the dialog fragment
does not have a view.

Test: added test in DialogFragmentTest
Fixes: 152043404
Change-Id: I524d0509cdd21ad31efee83d1e6002a45609549b