Manual drawing reordering
Originally we tried to use the auto redrawing based on RenderNode's elevation which proved to be tricky to support and caused a number of bugs. Instead of relaying on this mechanism we decided to decouple shadows size from z index as features and this CL is a first part of it. Here we still use DrawLayerModifier's elevation as a zIndex, but we do the drawing reordering manually.
This should also improve the drawing performance as we invalidate less(on apis < 29) and do expensive enableZ calls less often.

Fixes: 152600006
Fixes: 152696056
Bug: 151818355
Bug: 152417501
Test: new tests, manually
Change-Id: I45aa0e751b9d3bba74babcd25a835d8b34b99c24