Ensure the root graph is matched when deep linking
findInvalidDestinationDisplayNameInDeepLink()
is designed to ensure that each destination in
the deep link is found in the graph.

NavDestination.buildDeepLinkIds() always uses
the parent graph's ID when deep linking to the
start destination of a graph. This means that
the first ID in the deep link array is always the
ID of the root of the graph.

Rather than assume that the root graph always
matches, we should explicitly check that the first
ID matches the ID on the root graph. This is
particularly important in cases where there are
multiple NavHostFragments in the same activity
(either nested or side by side) to ensure that only
the NavHostFragment that has the graph being
deep linked into handles the deep link and the others
properly ignore it (instead of crashing).

Test: newly added test passes
BUG: 147378752
Change-Id: Ic69d684f8886f177c846ea1f62f6e8966b611c72