Preparation to remove sync methods from test rule.
Currently runOnIdleCompose requires ComposeTestRule. This dependency is
not neccessary and reduces the flexibility for developers to avoid the
need to use that rule. This CL introduces these methods as global
functions. Currently they are internal. Once this lands successfully I
will go ahead and make them public and remove them from ComposeTestRule.

Bug: b/152850851
Test: N/A
Change-Id: I421fa09122dcd0bc2d3ab52b08953e05c41b988d