Invalidate layer when draw modifier size changes
The CL invalidates the layer corresponding to a draw modifier when the
size of the draw modifier changes. This is required for the case when
there is a layer between the draw modifier and the layout node changing
size causing the draw modifier to change size, and the parent of the
layout node changing size is not changing size.

The bug was unsurfaced by If32acbfac08c677b80f9e4d5f624fe15c95ac60d.
Before that, there was a bug in DelegatingNodeWrapper which was calling
wrapped.place rather than performPlace, which was not setting dirtySize
to false on wrapped. This was causing unneeded redrawing for the layer
of the parent layout node even if this was not changing size. Setting
dirtySize correctly in the CL mentioned above should also be a
performance improvement by preventing unneeded redrawing.

Test: ran AndroidLayoutDrawTest
Bug: 152742986
Change-Id: I7df8d3b07b2c86ed499f6c047bc4775fda0c4eca