Implement Worker injection code generation
Similar to ViewModels, Worker are assisted injected and a Hilt
module for the multibinding map of assisted factories is generated.
To reused code assisted factory generator was created to be used by
both ViewModel and Worker generators.

This CL also adds a small integration app to demonstrate usage and to
further write tests.

A follow-up CL will contain validation.

Test: hilt-compiler:test
Change-Id: I6484aecb03b6e2877020755e8d280fd0f42c9df5