MediaRouter: Clean up callbacks in onDestroy
When MediaRouteProviderService deals with two different intents
onUnbind is not a good point to clean up.

Since the service will be destroyed when the last client has disconnected,
move the clean up logic into onDestory.

Bug: 153528451
Test: Sideload GmsCore with mediarouter.jar including this CL.
Support v7 demos should be launched before the sideloading to
simulate the behavior (onUnbind is called but onDestroy isn't called)

Change-Id: Ib4d78aca5f8128d8960034cb1b87d561d66531f8