API change: Support OutputFileOptions in CameraView#takePicture
To align the ImageCapture#takePicture API, modify the CameraView
API to support the OutputFileOptions input.
User can use URI or OptputStream if the android version only
support the scoped-storage.

Relnote: "Update the CameraView#takePicture() API to allow saving images to Uri and OutputStream. Update the test app to use Uri as the canonical example."

Bug: 153607583
Test: ./gradlew bOS &&
./gradlew camera:camera-core:connectedAndroidTest &&
./gradlew camera:camera-camera2:connectedAndroidTest &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:camera-core:test &&
./gradlew camera:camera-camera2:test &&
./gradlew camera:camera-extensions:test &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:integration-tests:camera-testapp-extensions:connectedAndroidTest

Change-Id: Ia2459465f2e960ab46409fceda367e5bc2678963