Migrate SemanticsPredicate to SemanticsMatcher.
Renames SemanticsPredicate to SemanticsMatcher but also changes the way
it filters out nodes. Instead of boolean condition it get a list of
nodes and return a filtered list of nodes. This enables to implement
boolean conditions like before but also enables to make projections of
the data by allowing APIs like node.parent.assertExists().

Bug: 154105179
Test: Already covered by existing ones.
Change-Id: I6cd86d59325a691b930c736b31d9a99b310d66f1