Do not allow only an empty action in NavDeepLinks
We currently allow for actions to be empty in NavDeepLinks and
NavDeepLinkRequests. Allowing empty actions at runtime will cause the
manifest merger to clear the default action and make the deep link
useless.

We should not allow empty actions when constructing a
NavDeepLink (through java or Kotlin DSL) or NavDeepLink request.

It is okay to add an empty action in xml since we can just treat it as
null in the NavInflater.

Test: added test
Bug: 154161310
Change-Id: I1097ed6fa215a11f324397a518c8aeb025363b27