Synchronize test in composeTestRule.setContent
ComposeTestRule.setContent already made an effort to be fully drawn
after it returned from the method, but it didn't work and it is now
handled by FirstDrawIdlingResource. All we need to do in setContent
is call waitForIdle().

This also adds waitForIdle() as a global method in Synchronization.

Bug: 154079800
Test: ./gradlew ui:ui-test:cC
Relnote: "No longer need to explicitly call onIdle() after
`ComposeTestRule.setContent {}`."

Change-Id: Iafb2e031362bb3cd179947642299f01f8c720b15