Ensure that markState() handles overridden getLifecycle()
markState() is responsible for updating the
Fragment's Lifecycle to CREATED when
onSaveInstanceState() is called. It does this
by checking the Lifecycle and, if it is at least
STARTED, moves it down to CREATED.

When a Fragment overrides getLifecycle(), it may
return a different object than the internal
mLifecycleRegistry that is used within Fragment.

By ensuring that we call getCurrentState() on
the same mLifecycleRegistry we're updating, we
can avoid infinite loops where the Lifecycle we're
checking is the same one we're updating.

Test: new FragmentArchLifecycleTest
Fixes: 154245967
Change-Id: Ice355a401d4dae799eb3ed447297ef814eff536f