Change NavDeepLink Uri APIs to UriPattern
Since NavDeepLink takes a string that is a uri pattern and not an actual
Uri, this changes all of the methods related to that pattern to say
uriPattern instead of just uri.

Also fixed the javaDoc in NavDeepLinkRequest that said the uri parameter
was a pattern to just say uri.

Test: all tests pass, ./gradlew checkApi
Bug: 154165119
Change-Id: I3f87edc56aee948b0220a098779e30ba5e9065e0