Update composer to calculate the node position correctly
Inserting, moving or deleted nodes from a child and a parent
node in the same composition could cause the composer to be
confused about which node is being updated. This could result
in an index out-of-boudns or nodes showing up or being removed
from the wrong place in the tree.

Fixes: b/154650546
Test: new test in ./gradlew :compose:compose-runtime:testDebugUnitTest
Change-Id: Ifcc5b3da8eff260930d420b3813b4811d7620482