Pass one master out directory and clear it before running
This fixes the issue where previous runs of this script
would persist, creating an out directory that was actually
wrong.

Secondly, this will eventually enable me to create a full
end to end test because we will need to specify a temp test
directory that the test can use and delete.

Also, remove an old function that should have been deleted
when we moved this script from buildSrc into it's own
python script in developement/release-notes/*

Test: ./generateReleaseNotes.py 1586970000000
Test: ./testReleaseNotes.py
Change-Id: Ia338abd8c59680199ca7d9fca0423bcbad7f5118