Fix measurement in width and height modifiers
Before this CL, the width and height modifiers were measuring their
wrapped layout with [0, infinity) constraints in the opposite axis. This
CL fixes this behavior, by keeping the opposite axis constraints the
modifier is measured with.

Relnote: Fixed a bug in the width and height modifiers that was causing the wrapped layout to be measured with no opposite axis constraints.
Bug: 154797971
Test: LayoutSizeTest
Change-Id: I210b84ad765c3ab33e593f027245ab135dd036f4